1、修复message打包之后不能运行的bug 2、完善各个模块功能 3、将商品评价表由goods模块移动到order模块

+++ irh-service/irh-service-goods/src/main/java/top/imuster/goods/dao/ProductRecommendDao.java
@@ -12,6 +12,5 @@ import top.imuster.goods.api.dto.ProductRecommendDto;
  */
 @Repository
 public interface ProductRecommendDao extends MongoRepository<ProductRecommendDto, String> {
-
-    ProductRecommendDto findByUserId(Integer userId);
+    ProductRecommendDto findByUserId(Long userId);
 }
